Solr compiled items feature added. First responses even faster.
