+++ dspace-oai/dspace-oai-api/src/main/java/org/dspace/xoai/exceptions/CompilingException.java
@@ -1,5 +1,6 @@
 package org.dspace.xoai.exceptions;
 
+@SuppressWarnings("serial")
 public class CompilingException extends Exception
 {
