+++ dspace-oai/dspace-oai-api/src/main/java/org/dspace/xoai/exceptions/MetadataBindException.java
@@ -1,0 +1,27 @@
+package org.dspace.xoai.exceptions;
+
+@SuppressWarnings("serial")
+public class MetadataBindException extends Exception
+{
+
+    public MetadataBindException()
+    {
+        super();
+    }
+
+    public MetadataBindException(String arg0, Throwable arg1)
+    {
+        super(arg0, arg1);
+    }
+
+    public MetadataBindException(String arg0)
+    {
+        super(arg0);
+    }
+
+    public MetadataBindException(Throwable arg0)
+    {
+        super(arg0);
+    }
+    
+}
