+++ dspace-oai/dspace-oai-api/src/main/java/org/dspace/xoai/filter/DSpaceMetadataExistsFilter.java
@@ -12,7 +12,7 @@ import java.sql.SQLException;
 import org.apache.log4j.LogManager;
 import org.apache.log4j.Logger;
 import org.dspace.core.Context;
-import org.dspace.xoai.data.DSpaceItem;
+import org.dspace.xoai.data.DSpaceDatabaseItem;
 import org.dspace.xoai.exceptions.InvalidMetadataFieldException;
 import org.dspace.xoai.util.MetadataFieldManager;
 
@@ -57,7 +57,7 @@ public class DSpaceMetadataExistsFilter extends DSpaceFilter
     }
 
     @Override
-    public boolean isShown(DSpaceItem item)
+    public boolean isShown(DSpaceDatabaseItem item)
     {
         if (item.getMetadata(this.getField()+".*").size() > 0)
             return true;
