+++ dspace-oai/dspace-oai-api/src/main/java/org/dspace/xoai/filter/DateUntilFilter.java
@@ -12,7 +12,7 @@ import java.util.Date;
 import org.apache.solr.client.solrj.util.ClientUtils;
 
 import org.dspace.core.Context;
-import org.dspace.xoai.data.DSpaceItem;
+import org.dspace.xoai.data.DSpaceDatabaseItem;
 
 /**
  * 
@@ -35,7 +35,7 @@ public class DateUntilFilter extends DSpaceFilter
     }
 
     @Override
-    public boolean isShown(DSpaceItem item)
+    public boolean isShown(DSpaceDatabaseItem item)
     {
         if (item.getDatestamp().compareTo(_date) <= 0)
             return true;
