+++ dspace/solr/oai/conf/schema.xml
@@ -152,6 +152,9 @@
    <field name="item.lastmodified" type="date" indexed="true" stored="true" multiValued="false" />
    <field name="item.submitter" type="string" indexed="true" stored="true" multiValued="false" />
    <field name="item.deleted" type="boolean" indexed="true" stored="true" multiValued="false" />
+   
+   <!-- Item compiled -->
+   <field name="item.compile" type="string" indexed="false" stored="true" multiValued="false" />
 
    <!-- Item metadata -->
    <dynamicField name="metadata.*" type="string" indexed="true" stored="true" multiValued="true" />
